Добавил авторизацию через мобильную версию приложения

+++ AccountServer/AccountData/Account/Entities/SocialDb.cs
@@ -8,14 +8,14 @@ namespace AccountData.Account.Entities
   public class SocialDb
   {
     [Key]
-    public int SocialId { get; set; }
+    public long SocialId { get; set; }
 
     public SocialType SocialType { get; set; }
-    public int InternalId { get; set; }
+    public long InternalId { get; set; }
     public long ExternalId { get; set; }
     public string Email { get; set; }
     public string PhoneNumber { get; set; }
     public string ExternalToken { get; set; }
-    public int ExpiresIn { get; set; }
+    public long ExpiresIn { get; set; }
   }
 }
